+++ back-end/app/Http/Controllers/LessonController.php
@@ -16,7 +16,13 @@ class LessonController extends Controller
         $this->middleware('auth:api',['except' => ['getAllLessons','getOneLesson','addNewLesson','updateLesson','deleteLesson','changeStatusLesson']]);
     }
     public function getAllLessons(Request $request){
-        $lessons = Lesson::all();
+        $login = auth()->user();
+        if($login && $login->is_admin == true){
+            $lessons = Lesson::all();
+        }
+        else{
+            $lessons = Lesson::where('status','Active')->get();
+        }
         return response()->json([
             'lessons'=>$lessons
         ], 200);
@@ -29,14 +35,20 @@ class LessonController extends Controller
             return response()->json(['error'=>$validator->errors()], 400);      
         }
 
-        $lesson = Lesson::find($request->id);
+        $login = auth()->user();
+        if($login && $login->is_admin == true){
+            $lessons = Lesson::find($request->id);
+        }
+        else{
+            $lessons = Lesson::where('status','Active')->where('id',$request->id)->get();
+        }
         return response()->json([
             'lesson'=>$lesson
         ], 200);
     }
     public function addNewLesson(Request $request){
         $login = auth()->user();
-        if($login->is_admin == true){
+        if($login && $login->is_admin == true){
             $validator = Validator::make($request->all(), [
                 'name' => 'required|min:1|max:255|unique:lesson,name',
                 'content_id'=>'required|exists:content,id',
@@ -85,18 +97,32 @@ class LessonController extends Controller
     }
     public function updateLesson(Request $request){
         $login = auth()->user();
-        if($login->is_admin == true){
+        if($login && $login && $login->is_admin == true){
             $validator = Validator::make($request->all(), [
                 'id' => 'required|exists:lesson,id',
-                'name' => 'required|min:1|max:255',
-                'content_id'=>'required|exists:content,id',
-                'file'=>'required|mimetypes:video/avi,video/mpeg,video/mp4',
-                'path'=>'required',
-                'status'=>'required|in:Active,Block',
-                'description'=>'required',
+                'name' => 'max:255|string',
+                'content_id'=>'exists:content,id',
+                'file'=>'mimetypes:video/avi,video/mpeg,video/mp4',
+                'path'=>'',
+                'status'=>'in:Active,Block',
+                'description'=>'',
             ]);
             if ($validator->fails()) {
                 return response()->json(['error'=>$validator->errors()], 400);      
+            }
+           
+            $lesson = Lesson::find($request->id);
+            if($lesson->name == $request->name || $request->name == null){
+                $lesson->name = $lesson->name;
+            }
+            else{
+                $validator = Validator::make($request->all(), [
+                    'name' => 'unique:lesson,name',
+                ]);
+                if ($validator->fails()) {
+                    return response()->json(['error'=>$validator->errors()], 400);      
+                }
+                $lesson->name = $request->name;
             }
             if($request->hasfile('file')) {
                 $destinationPath = public_path().DIRECTORY_SEPARATOR.'upload'.DIRECTORY_SEPARATOR.'lesson_files';
@@ -111,25 +137,17 @@ class LessonController extends Controller
                 $file->move(public_path().DIRECTORY_SEPARATOR.'upload'.DIRECTORY_SEPARATOR.'lesson_files', $newFileName);
                 $linkFile = $request->getSchemeAndHttpHost().'/'.'upload'.'/'.'lesson_files'.'/'.$newFileName;
             }
-            $lesson = Lesson::find($request->id);
-            if($lesson->name == $request->name){
-                $lesson->name = $request->name;
+            $request->content_id == null ? $lesson->content_id = $lesson->content_id : $lesson->content_id = $request->content_id;
+            if($request->hasfile('file')){
+                File::delete($destinationPath.'/'.$lesson->file_name);
+                $lesson->file_name = $newFileName;
             }
             else{
-                $validator = Validator::make($request->all(), [
-                    'name' => 'unique:lesson,name',
-                ]);
-                if ($validator->fails()) {
-                    return response()->json(['error'=>$validator->errors()], 400);      
-                }
-                $lesson->name = $request->name;
+                $lesson->file_name = $lesson->file_name;
             }
-            $lesson->content_id = $request->content_id;
-            File::delete($destinationPath.'/'.$lesson->file_name);
-            $lesson->file_name = $newFileName;
-            $lesson->path = $request -> path;
-            $lesson->status = $request->status;
-            $lesson->description = $request->description;
+            $request->path == null ? $lesson->path = $lesson -> path : $lesson->path = $request -> path;
+            $request->status == null ? $lesson->status = $lesson->status : $lesson->status = $request->status;
+            $request->description == null ? $lesson->description = $lesson->description : $lesson->description = $request->description;
             $lesson->updated_at = Carbon::now('Asia/Ho_Chi_Minh');
             $lesson->save();
             return response()->json([
@@ -146,7 +164,7 @@ class LessonController extends Controller
     }
     public function deleteLesson(Request $request){
         $login = auth()->user();
-        if($login->is_admin == true){
+        if($login && $login->is_admin == true){
             $validator = Validator::make($request->all(), [
                 'id' => 'required|exists:lesson,id',
             ]);
@@ -171,7 +189,7 @@ class LessonController extends Controller
     }
     public function changeStatusLesson(Request $request){
         $login = auth()->user();
-        if($login->is_admin == true){
+        if($login && $login->is_admin == true){
             $validator = Validator::make($request->all(), [
                 'id' => 'required|exists:lesson,id',
             ]);
